Refactoed code for directory structure
