--- lib/features/counter/chart/chart.dart
@@ -1,2 +1,0 @@
-export 'chart_page.dart';
-export 'heatmap_page.dart';
