+++ lib/features/counter/counter_details_page.dart
@@ -2,8 +2,7 @@ import 'package:flutter/material.dart';
 
 import 'package:provider/provider.dart';
 
-import '../chart/base_heatmap_widget.dart';
-import 'chart/chart.dart';
+import '../chart/chart.dart';
 import 'model/model.dart';
 import 'provider/counter_list_provider.dart';
 
@@ -156,5 +155,80 @@ class CounterDetailsPage extends StatelessWidget {
         ),
       ),
     );
+  }
+}
+
+class HeatmapPage extends StatelessWidget {
+  final CounterModel counter;
+
+  const HeatmapPage({super.key, required this.counter});
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.all(20),
+      child: BaseHeatmapWidget(
+        heatmapData: Provider.of<CounterListProvider>(context)
+            .extractCountsByDay(counter),
+      ),
+    );
+  }
+}
+
+class ChartPage extends StatelessWidget {
+  final CounterModel counter;
+
+  const ChartPage({super.key, required this.counter});
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.all(16.0),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.stretch,
+        children: <Widget>[
+          const Text(
+            'Exercise Counts - Last 7 Days',
+            style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
+            textAlign: TextAlign.center,
+          ),
+          const SizedBox(height: 20),
+          Consumer<CounterChartProvider>(
+            builder: (context, chartProvider, child) {
+              final barGroups = chartProvider.processDataForChart(counter);
+              final maxY = barGroups
+                      .map((group) => group.barRods.first.toY)
+                      .fold(0.0, (max, y) => y > max ? y : max) +
+                  5;
+              return !chartProvider.isProcessed
+                  ? const CircularProgressIndicator()
+                  : BaseChartWidget(
+                      barGroups: barGroups,
+                      dayLabel: (value) => chartProvider.getDayOfWeek(value),
+                      yAxisInterval: calculateYAxisInterval(maxY),
+                    );
+            },
+          ),
+          const SizedBox(height: 20),
+          const Text(
+            'Days are based on the last 7 days.',
+            style: TextStyle(fontSize: 12, color: Colors.grey),
+            textAlign: TextAlign.center,
+          ),
+        ],
+      ),
+    );
+  }
+
+  static double calculateYAxisInterval(double maxY) {
+    if (maxY <= 10) {
+      return 1;
+    } else if (maxY <= 50) {
+      return 5;
+    } else if (maxY <= 100) {
+      return 10;
+    } else {
+      return 20; // Or some other logic for larger maxY values
+    }
   }
 }
