+++ lib/features/timer/model/duration_interval.dart
