+++ lib/features/timer/model/model.dart
@@ -1,2 +1,3 @@
+export 'duration_interval.dart';
 export 'flat_timer_model.dart';
 export 'timer_model.dart';
